Move navigation menus REST client out of FluxC
Move NavMenuRestClient from FluxC to WordPress module to align with
the goal of phasing out FluxC in favor of wordpress-rs.

- Move REST client to WordPress/ui/navmenus/data/
- Rename NavMenuRsRestClient to NavMenuRestClient
- Delete unused NavMenuStore and NavMenuAction from FluxC
- Keep data models in FluxC for now (needed for WellSql persistence)
